requirements.txt
Flask==2.2.2
numpy==1.21.6
yfinance==0.2.20
urllib3==2.0.2
pandas==1.3.5
ta==0.10.2
flask-cors==3.0.10
Selenium==4.10.0
python-dotenv==0.21.1